backend/goals-maker/pyproject.toml
[tool.poetry]
name = "goals-maker"
version = "0.1.0"
description = ""
authors = ["Muhammad Hamza <hamzaadil56@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
langchain-groq = "^0.2.1"
langchain = "^0.3.8"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
